+++ app/src/main/java/com/pmydm/projecterato/SettingActivity.kt
@@ -2,21 +2,19 @@ package com.pmydm.projecterato
 
 import android.content.Intent
 import android.os.Bundle
+import android.widget.Button
 import android.widget.ImageButton
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 
-class SettingsActivity : AppCompatActivity() {
+class SettingActivity : AppCompatActivity() {
     private var volumen=true
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
+
         val imageButtonVolver: ImageButton = findViewById(R.id.imageButtonVolver)
 
         imageButtonVolver.setOnClickListener {
-            val intent = Intent(this, MainActivity::class.java)
+            val intent = Intent(this, ProfileMenuNLActivity::class.java)
             startActivity(intent)
         }
 
@@ -29,7 +27,16 @@ class SettingsActivity : AppCompatActivity() {
             } else {
                 imageButtonVolumen.setImageResource(R.drawable.iconovolumenencendido)
                 volumen=true
-            }  // Alternar el estado
+            }
+            // Alternar el estado
+        }
+
+        val buttonCerrarSesion : Button = findViewById(R.id.buttonCerrarSesion)
+
+        buttonCerrarSesion.setOnClickListener {
+            val intent = Intent(this, ProfileMenuNLActivity::class.java)
+            MainActivity.setLogin()
+            startActivity(intent)
         }
     }
 }
